+++ dashboard/news/delete.php
@@ -2,18 +2,26 @@
     include_once($_SERVER['DOCUMENT_ROOT'] . '/dashboard/header.php');
     $user = new User($conn);
 
-    if(isset($_GET['id'])) {
-        $id = $_GET['id'];
+    //Controleer als de gebruiker de rol Eigenaar of administrator heeft.
+    if($user->has_role('Eigenaar') || $user->has_role('Administrator')) {
+        //Controleer als er een ID als GET parameter in de URL wordt meegegeven.
+        if (isset($_GET['id'])) {
+            $id = $_GET['id'];
 
-        $newsitem = $conn ->prepare("SELECT image FROM news WHERE ID = " . $id);
-        $newsitem->execute();
-        $newsitemimage = $newsitem->fetch(PDO::FETCH_ASSOC);
+            //Haal de afbeelding op van het nieuwsbericht.
+            $newsitem = $conn->prepare("SELECT image FROM news WHERE ID = " . $id);
+            $newsitem->execute();
+            $newsitemimage = $newsitem->fetch(PDO::FETCH_ASSOC);
 
-        $stmt = $conn->prepare("DELETE FROM news WHERE ID = " . $id);
-        $stmt->execute();
-        unlink($_SERVER['DOCUMENT_ROOT'] . '/assets/images/news/' . $newsitemimage['image']);
+            //Verwijder nieuwsbericht uit de database en verwijder de afbeelding van de server.
+            $stmt = $conn->prepare("DELETE FROM news WHERE ID = " . $id);
+            $stmt->execute();
+            unlink($_SERVER['DOCUMENT_ROOT'] . '/assets/images/news/' . $newsitemimage['image']);
 
-        $user->redirect('/dashboard/news');
+            $user->redirect('/dashboard/news');
+        } else {
+            echo 'Product niet gevonden';
+        }
     } else {
-        echo 'Product niet gevonden';
+        $user->redirect('/dashboard');
     }
