+++ dashboard/reviews/delete.php
@@ -1,9 +1,22 @@
 <?php
     include_once($_SERVER['DOCUMENT_ROOT'] . '/dashboard/header.php');
-    $url = "/";
-    $id = $_GET["id"];
-        $delete_reviews = $conn->prepare('DELETE FROM reviews WHERE ID = :reviewid');
-        $delete_reviews -> execute(array(
-            ':reviewid' => $id));
-        header('Location: ' . $_SERVER['HTTP_REFERER']);
+    $user = new User($conn);
+
+    //Controleer als de gebruiker de rol Eigenaar, administrator of medewerker heeft.
+    if($user->has_role('Eigenaar') || $user->has_role('Administrator') || $user->has_role('Medewerker')) {
+        //Controleer als er een ID als GET parameter wordt meegegeven aan de URL.
+        if (isset($_GET['id'])) {
+            $id = $_GET["id"];
+
+            //Verwijder de opgevraagde review.
+            $delete_reviews = $conn->prepare('DELETE FROM reviews WHERE ID = :reviewid');
+            $delete_reviews->execute(array(
+                ':reviewid' => $id));
+            $user->redirect('/dashboard/reviews');
+        } else {
+            echo 'Review kon niet worden gevonden';
+        }
+    } else {
+        $user->redirect('/dashboard');
+    }
 ?>
